// src/constant/enums.ts
export enum ValueType {
  NULL = '[object Null]',
  UNDEFINED = '[object Undefined]',
  FUNCTION = '[object Function]',
  STRING = '[object String]',
  NUMBER = '[object Number]',
  SYMBOL = '[object Symbol]',
  BIGINT = '[object BigInt]',
  OBJECT = '[object Object]',
  ARRAY = '[object Array]',
  REGEXP = '[object RegExp]',
  BLOB = '[object Blob]'
}

export enum PasswordStrenthENUM {
  FEEBLE = 'feeble',
  MIDDLE = 'middle',
  BETTER = 'better',
  FORTISSIMO = 'fortissimo'
}

export enum RedirectionType {
  APIURL = 'apiurl',
  DIREACTIVEURL = 'directiveurl',
  ENTRYURL = 'entryurl',
  PRINTURL = 'printurl',
  MICROURL = 'microurl'
}

export enum BooleanType {
  TRUE = 1,
  FALSE = 0
}

export enum BuildType {
  VUECLI = 'vue-cli',
  VITE = 'vite'
}